Fix an issue where 'Backup Notes' crashes the app, Backup Notes not working yet.
Backup notes doesn't work yet. I am getting an ActivityNotFound exception. At first, thought it was an issue of not
having the appropriate app installed to handle the intent. The 'Backup Notes' function sends a txt file as an intent,
but within my emulator I didn't have any apps installed to handle the txt file on the other end. However, I tried it on
a device with Dropbox installed and I'm still getting the same result. Will look into this further.